Split Host and Deployer class into different files.
