+++ app/deployer/deployer.py
@@ -1,4 +1,5 @@
 import os
+from .host import Host
 from python_terraform import Terraform
 from flask import render_template
 from jinja2 import TemplateNotFound
@@ -85,16 +86,3 @@ class Deployer:
                                                                         stderr
                                                                         )
                     )
-
-
-class Host:
-    '''
-    Contains all variables introduced by the user used to create a VM in Azure.
-    '''
-    def __init__(self, name=None, dnsname=None, username=None, passwd=None,
-                 public_key=None):
-        self.name = name
-        self.dnsname = dnsname
-        self.username = username
-        self.passwd = passwd
-        self.public_key = public_key
